# swagger.json
{"openapi":"3.0.0","paths":{"/api/auth/send-sms":{"post":{"operationId":"AuthController_sendSms","summary":"发送短信验证码","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SendSmsDto"}}}},"responses":{"201":{"description":""}},"tags":["认证"]}},"/api/auth/login-phone":{"post":{"operationId":"AuthController_loginPhone","summary":"手机号验证码登录","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginPhoneDto"}}}},"responses":{"201":{"description":""}},"tags":["认证"]}},"/api/auth/refresh":{"post":{"operationId":"AuthController_refreshToken","summary":"刷新Token","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenDto"}}}},"responses":{"201":{"description":""}},"tags":["认证"]}},"/api/user/profile":{"get":{"operationId":"UserController_getProfile","summary":"获取个人资料","parameters":[],"responses":{"200":{"description":""}},"tags":["用户"],"security":[{"bearer":[]}]},"put":{"operationId":"UserController_updateProfile","summary":"更新个人资料","parameters":[],"responses":{"200":{"description":""}},"tags":["用户"],"security":[{"bearer":[]}]}},"/api/notes":{"post":{"operationId":"NoteController_create","summary":"发布笔记","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateNoteDto"}}}},"responses":{"201":{"description":""}},"tags":["笔记"],"security":[{"bearer":[]}]},"get":{"operationId":"NoteController_findAll","summary":"笔记列表","parameters":[{"name":"mine","required":true,"in":"query","schema":{"type":"string"}},{"name":"page","required":true,"in":"query","schema":{"type":"string"}},{"name":"limit","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["笔记"],"security":[{"bearer":[]}]}},"/api/notes/{id}":{"get":{"operationId":"NoteController_findOne","summary":"笔记详情","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["笔记"],"security":[{"bearer":[]}]},"put":{"operationId":"NoteController_update","summary":"编辑笔记","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["笔记"],"security":[{"bearer":[]}]},"delete":{"operationId":"NoteController_remove","summary":"删除笔记","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["笔记"],"security":[{"bearer":[]}]}},"/api/pay/wechat-unified":{"post":{"operationId":"PaymentController_wechatUnified","summary":"微信统一下单","parameters":[],"responses":{"201":{"description":""}},"tags":["支付"],"security":[{"bearer":[]}]}},"/api/pay/alipay-create":{"post":{"operationId":"PaymentController_alipayCreate","summary":"支付宝下单","parameters":[],"responses":{"201":{"description":""}},"tags":["支付"],"security":[{"bearer":[]}]}},"/api/pay/notify/wechat":{"post":{"operationId":"PaymentController_wechatNotify","summary":"微信支付回调","parameters":[],"responses":{"201":{"description":""}},"tags":["支付"],"security":[{"bearer":[]}]}},"/api/pay/notify/alipay":{"post":{"operationId":"PaymentController_alipayNotify","summary":"支付宝支付回调","parameters":[],"responses":{"201":{"description":""}},"tags":["支付"],"security":[{"bearer":[]}]}},"/api/pay/purchase-note":{"post":{"operationId":"PaymentController_purchaseNote","summary":"购买笔记","parameters":[],"responses":{"201":{"description":""}},"tags":["支付"],"security":[{"bearer":[]}]}},"/api/iap/create":{"post":{"operationId":"IapController_createIap","summary":"创建IAP订单","parameters":[],"responses":{"201":{"description":""}},"tags":["Apple IAP"],"security":[{"bearer":[]}]}},"/api/iap/verify":{"post":{"operationId":"IapController_verifyIap","summary":"验证IAP收据","parameters":[],"responses":{"201":{"description":""}},"tags":["Apple IAP"],"security":[{"bearer":[]}]}},"/api/withdrawals":{"post":{"operationId":"WithdrawalController_create","summary":"申请提现","parameters":[],"responses":{"201":{"description":""}},"tags":["提现"],"security":[{"bearer":[]}]},"get":{"operationId":"WithdrawalController_findByUser","summary":"我的提现记录","parameters":[{"name":"page","required":true,"in":"query","schema":{"type":"string"}},{"name":"limit","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["提现"],"security":[{"bearer":[]}]}},"/api/admin/auth/login":{"post":{"operationId":"AdminController_login","summary":"管理员登录","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginAdminDto"}}}},"responses":{"201":{"description":""}},"tags":["管理后台"]}},"/api/admin/dashboard":{"get":{"operationId":"AdminController_getDashboard","summary":"统计数据","parameters":[],"responses":{"200":{"description":""}},"tags":["管理后台"],"security":[{"bearer":[]}]}},"/api/admin/users":{"get":{"operationId":"AdminController_findUsers","summary":"用户列表","parameters":[],"responses":{"200":{"description":""}},"tags":["管理后台"],"security":[{"bearer":[]}]}},"/api/admin/users/{id}/status":{"put":{"operationId":"AdminController_updateUserStatus","summary":"更新用户状态","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["管理后台"],"security":[{"bearer":[]}]}},"/api/admin/notes":{"get":{"operationId":"AdminController_findNotes","summary":"笔记列表","parameters":[],"responses":{"200":{"description":""}},"tags":["管理后台"],"security":[{"bearer":[]}]}},"/api/admin/notes/{id}/status":{"put":{"operationId":"AdminController_updateNoteStatus","summary":"更新笔记状态","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["管理后台"],"security":[{"bearer":[]}]}},"/api/admin/orders":{"get":{"operationId":"AdminController_findOrders","summary":"订单列表","parameters":[],"responses":{"200":{"description":""}},"tags":["管理后台"],"security":[{"bearer":[]}]}},"/api/admin/withdrawals":{"get":{"operationId":"AdminController_findWithdrawals","summary":"提现列表","parameters":[],"responses":{"200":{"description":""}},"tags":["管理后台"],"security":[{"bearer":[]}]}},"/api/admin/withdrawals/{id}":{"put":{"operationId":"AdminController_reviewWithdrawal","summary":"审核提现","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["管理后台"],"security":[{"bearer":[]}]}}},"info":{"title":"心事盒 API","description":"心事盒APP后端接口文档","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"SendSmsDto":{"type":"object","properties":{}},"LoginPhoneDto":{"type":"object","properties":{}},"RefreshTokenDto":{"type":"object","properties":{}},"CreateNoteDto":{"type":"object","properties":{}},"LoginAdminDto":{"type":"object","properties":{}}}}}